Design: 메인페이지에서 채널색상에 따라 tailwind 적용을 style로 변경
